just inline it
